feat: add admin pages

+++ backend/src/main/java/com/harry/clio/service/impl/CategoryServiceImpl.java
@@ -2,24 +2,31 @@ package com.harry.clio.service.impl;
 
 import com.harry.clio.dto.category.CategoryResponse;
 import com.harry.clio.dto.category.CreateCategoryRequest;
+import com.harry.clio.exception.BadRequestException;
 import com.harry.clio.exception.ResourceNotFoundException;
 import com.harry.clio.mapper.CategoryMapper;
 import com.harry.clio.model.Category;
+import com.harry.clio.repository.BookRepository;
 import com.harry.clio.repository.CategoryRepository;
 import com.harry.clio.service.CategoryService;
 
 import lombok.RequiredArgsConstructor;
 
+import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.cache.annotation.Cacheable;
+import org.springframework.cache.annotation.Caching;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-@RequiredArgsConstructor
 @Service
+@RequiredArgsConstructor
 public class CategoryServiceImpl implements CategoryService {
     private final CategoryRepository categoryRepository;
     private final CategoryMapper categoryMapper;
+    private final BookRepository bookRepository;
 
     @Cacheable(value = "categories")
     public List<CategoryResponse> getCategories() {
@@ -31,10 +38,11 @@ public class CategoryServiceImpl implements CategoryService {
     private Category getCategoryOrThrow(int categoryId) {
         return categoryRepository
                 .findById(categoryId)
-                .orElseThrow(() -> new ResourceNotFoundException("Không tìm thấy danh mục"));
+                .orElseThrow(() -> new ResourceNotFoundException("Không tìm thấy thể loại"));
     }
 
     @Override
+    @CacheEvict(cacheNames = "categories", allEntries = true)
     public CategoryResponse getCategoryById(int categoryId) {
         return categoryMapper.toResponse(getCategoryOrThrow(categoryId));
     }
@@ -46,6 +54,12 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
+    @Caching(
+            evict = {
+                @CacheEvict(cacheNames = "categories", allEntries = true),
+                @CacheEvict(cacheNames = "books", allEntries = true),
+                @CacheEvict(cacheNames = "book-detail", allEntries = true)
+            })
     public CategoryResponse updateCategory(int categoryId, CreateCategoryRequest request) {
         Category category = getCategoryOrThrow(categoryId);
         category.setName(request.name());
@@ -53,8 +67,17 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
+    @CacheEvict(cacheNames = "categories", allEntries = true)
     public void deleteCategory(int categoryId) {
+        if (bookRepository.existsByCategoriesId(categoryId)) {
+            throw new BadRequestException("Không thể xóa do có sách đang trong thể loại");
+        }
         Category category = getCategoryOrThrow(categoryId);
         categoryRepository.delete(category);
+    }
+
+    @Override
+    public Page<CategoryResponse> getAdminCategories(Pageable pageable) {
+        return categoryRepository.findAll(pageable).map(categoryMapper::toResponse);
     }
 }
